Update README.md
My name is Roya! I am a first-generation student, an immigrant, and a Muslim woman of color. I am passionate about activism, climate reform, sustainability, education, and tech for good. I am also an avid reader, film nerd, and a museum addict. I hope to make an impact in communities such as women in tech and STEM education for all.